Refactoring: DataDict moves in separate files, utility functions move from BindingBase to DataHelper
