+++ bindings/Sofa/src/SofaPython3/Sofa/Core/Data/Binding_DataDict.cpp
@@ -1,0 +1,6 @@
+#include "Binding_DataDict.h"
+
+Binding_DataDict::Binding_DataDict()
+{
+
+}
