+++ bindings/Sofa/src/SofaPython3/Sofa/Core/Data/Binding_DataDict.h
@@ -1,0 +1,11 @@
+#ifndef BINDING_DATADICT_H
+#define BINDING_DATADICT_H
+
+
+class Binding_DataDict
+{
+public:
+    Binding_DataDict();
+};
+
+#endif // BINDING_DATADICT_H
