+++ bindings/Sofa/src/SofaPython3/Sofa/Core/Data/binding_datadict_doc.h
@@ -1,0 +1,4 @@
+#ifndef BINDING_DATADICT_DOC_H
+#define BINDING_DATADICT_DOC_H
+
+#endif // BINDING_DATADICT_DOC_H
